modified:   admin/admin.php 	modified:   admin/cetak.php 	modified:   admin/cetak_detail.php 	modified:   admin/data_kasir.php 	modified:   admin/data_paket.php 	modified:   admin/data_pelanggan.php 	modified:   admin/detail.php 	new file:   admin/head_css.php 	modified:   admin/laporan_transaksi.php 	new file:   admin/profile_admin.php 	modified:   admin/riwayat_transaksi.php 	modified:   admin/sidebar.php 	modified:   db.php 	modified:   index.php 	modified:   kasir/cetak_detail.php 	modified:   kasir/data_pelanggan.php 	modified:   kasir/detail.php 	new file:   kasir/head_css.php 	modified:   kasir/kasir.php 	modified:   kasir/modal_edit_detail.php 	modified:   kasir/profile_kasir.php 	modified:   kasir/riwayat_transaksi.php 	modified:   kasir/sidebar.php 	modified:   kasir/transaksi.php 	deleted:    laundry.sql 	new file:   pelanggan/cetak_detail.php 	new file:   pelanggan/detail.php 	new file:   pelanggan/head_css.php 	new file:   pelanggan/pelanggan.php 	new file:   pelanggan/profile_pelanggan.php 	new file:   pelanggan/riwayat_transaksi.php 	new file:   pelanggan/sidebar.php 	new file:   pelanggan/transaksi.php

+++ admin/data_kasir.php
@@ -5,7 +5,8 @@ $page = "data_kasir";
 if (empty($_SESSION['loginadmin'])) {
   header("Location: ../index.php");
 }
-$queryUser = "SELECT * FROM user";
+$username = $_SESSION['username'];
+$queryUser = "SELECT * FROM user WHERE username != '$username' ORDER BY id DESC";
 $execUser = mysqli_query($conn, $queryUser);
 $dataUser = mysqli_fetch_all($execUser, MYSQLI_ASSOC);
 if (isset($_GET['delete'])) {
@@ -15,21 +16,24 @@ if (isset($_GET['delete'])) {
   if ($execDeleteUser) {
     $log = $_SESSION['nama'] . "  " . "(" . $_SESSION['role'] . ")" . "  " . "Telah Menghapus user. Dengan id user (" . $id . "). Pada Data Kasir.";
     logger($log, "../../../../../");
-    header("location:data_kasir.php");
+    echo "<script>alert('User berhasil dihapus'); window.location.href='data_kasir.php';</script>";
+  } else {
+    echo "<script>alert('User gagal dihapus'); window.location.href='data_kasir.php';</script>";
   }
 }
 if (isset($_POST['edit'])) {
   $id = $_POST['id'];
   $nama = $_POST['nama'];
   $username = $_POST['username'];
-  $password = $_POST['password'];
   $rolee = $_POST['role'];
-  $queryUpdateUser = "UPDATE user SET `nama` = '$nama',`username` = '$username', `password`='$password',`role` = '$rolee' WHERE `user`.`id` = $id";
+  $queryUpdateUser = "UPDATE user SET `nama` = '$nama',`username` = '$username',`role` = '$rolee' WHERE `user`.`id` = $id";
   $execUpdateUser = mysqli_query($conn, $queryUpdateUser);
   if ($execUpdateUser) {
     $log = $_SESSION['nama'] . "  " . "(" . $_SESSION['role'] . ")" . "  " . "Telah Mengubah user. Dengan id user (" . $id . "). Pada Data Kasir.";
     logger($log, "../../../../../");
-    header("location:data_kasir.php");
+    echo "<script>alert('User berhasil diupdate'); window.location.href='data_kasir.php';</script>";
+  } else {
+    echo "<script>alert('User gagal diupdate'); window.location.href='data_kasir.php';</script>";
   }
 }
 if (isset($_POST['simpan'])) {
@@ -38,7 +42,7 @@ if (isset($_POST['simpan'])) {
   $dataTambah = mysqli_fetch_array($execTambah, MYSQLI_ASSOC);
   $nama = $_POST['nama'];
   $username = $_POST['username'];
-  $password = $_POST['password'];
+  $password = password_hash(1234, PASSWORD_ARGON2I);
   $rolee = $_POST['role'];
   $insert = "INSERT INTO `user` (`id`, `nama`, `username`, `password`, `role`) VALUES (NULL, '$nama', '$username', '$password', '$rolee');";
   // var_dump($kode);
@@ -47,26 +51,31 @@ if (isset($_POST['simpan'])) {
   if ($sql) {
     $log = $_SESSION['nama'] . "  " . "(" . $_SESSION['role'] . ")" . "  " . "Telah Menambahkan user. Dengan nama user '" . $nama . "' . Pada Data Kasir.";
     logger($log, "../../../../../");
-    header("location:data_kasir.php");
+    echo "<script>alert('User berhasil ditambahkan'); window.location.href='data_kasir.php';</script>";
+  } else {
+    echo "<script>alert('User gagal ditambahkan'); window.location.href='data_kasir.php';</script>";
+  }
+}
+if (isset($_POST['setPassDef'])) {
+  $id = $_POST['id'];
+  $passwordDefault = password_hash(1234, PASSWORD_ARGON2I);
+  $queryUpdateUser = "UPDATE user SET `password` = '$passwordDefault' WHERE `user`.`id` = $id";
+  $execUpdateUser = mysqli_query($conn, $queryUpdateUser);
+  if ($execUpdateUser) {
+    $log = $_SESSION['nama'] . "  " . "(" . $_SESSION['role'] . ")" . "  " . "Telah Men-setting password ke default. Dengan id user (" . $id . "). Pada Data Kasir.";
+    logger($log, "../../../../../");
+    echo "<script>alert('Password berhasil diatur ulang ke default'); window.location.href='data_kasir.php';</script>";
+  } else {
+    echo "<script>alert('Gagal mengatur ulang password'); window.location.href='data_kasir.php';</script>";
   }
 }
 ?>
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="stylesheet" href="../assets/css/main/app.css">
-  <link rel="stylesheet" href="../assets/css/main/app-dark.css">
-  <link rel="shortcut icon" href="../assets/images/logo/favicon.svg" type="image/x-icon">
-  <link rel="shortcut icon" href="../assets/images/logo/favicon.png" type="image/png">
-  <link rel="stylesheet" href="../assets/extensions/simple-datatables/style.css">
-  <link rel="stylesheet" href="../assets/css/pages/simple-datatables.css">
+  <?php include "head_css.php"; ?>
   <title>Data Kasir</title>
 </head>
-
 <body class="theme-dark" style="overflow-y: auto;">
   <div id="app">
     <?php include "sidebar.php" ?>
@@ -88,10 +97,8 @@ if (isset($_POST['simpan'])) {
                         <thead class="thead-dark">
                           <tr>
                             <th>No</th>
-                            <th>ID Anggota</th>
                             <th>Nama</th>
                             <th>Username</th>
-                            <th>Password</th>
                             <th>Role</th>
                             <th>Aksi</th>
                           </tr>
@@ -102,17 +109,12 @@ if (isset($_POST['simpan'])) {
                             <?php $no++ ?>
                             <tr>
                               <td><?= $no ?></td>
-                              <td><?= $user['id'] ?></td>
                               <td><?= $user['nama'] ?></td>
                               <td><?= $user['username'] ?></td>
-                              <td><?= $user['password'] ?></td>
                               <td><?= $user['role'] ?></td>
                               <td>
                                 <a href="" class="btn icon icon-left btn-primary" type="button" data-bs-toggle="modal" data-bs-target="#normal<?= $user['id']; ?>">
-                                  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-edit">
-                                    <path d="M11 4H4a2 2 0 0 0-2 2v14a2 2 0 0 0 2 2h14a2 2 0 0 0 2-2v-7"></path>
-                                    <path d="M18.5 2.5a2.121 2.121 0 0 1 3 3L12 15l-4 1 1-4 9.5-9.5z"></path>
-                                  </svg> UPDATE</a>
+                                <i class="bi bi-pencil-square"></i>UPDATE</a>
                                 <div class="modal fade text-left" id="normal<?= $user['id']; ?>" tabindex="-1" role="dialog" aria-labelledby="myModalLabel33" aria-hidden="true">
                                   <div class="modal-dialog modal-dialog-centered modal-dialog-scrollable" role="document">
                                     <div class="modal-content">
@@ -122,7 +124,6 @@ if (isset($_POST['simpan'])) {
                                           <i data-feather="x"></i>
                                         </button>
                                       </div>
-
                                       <form action="" method="post">
                                         <div class="modal-body">
                                           <label>Nama : </label>
@@ -133,15 +134,10 @@ if (isset($_POST['simpan'])) {
                                           <div class=" form-group">
                                             <input type="text" placeholder="Username" class="form-control" name="username" value="<?= $user['username'] ?>">
                                           </div>
-                                          <label>Password: </label>
-                                          <div class=" form-group">
-                                            <input type="password" placeholder="Password" class="form-control" name="password" value="<?= $user['password'] ?>">
-                                          </div>
                                           <label>Role: </label>
                                           <div class=" form-group">
                                             <select class="form-select" id="basicSelect" name="role">
                                               <option value="<?= $user['role'] ?>"><?= $user['role'] ?></option>
-                                              <!-- <option value="kasir">Kasir</option> -->
                                             </select>
                                           </div>
                                         </div>
@@ -151,16 +147,20 @@ if (isset($_POST['simpan'])) {
                                             <span class="d-none d-sm-block">Close</span>
                                           </button>
                                           <button type="submit" class="btn btn-primary ml-1" data-bs-dismiss="modal" name="edit" value="EDIT DATA">
-                                            <input type="text" class="visually-hidden" value="<?= $user['id'] ?>" name="id">
                                             <i class="bx bx-check d-block d-sm-none"></i>
                                             <span class="d-none d-sm-block">Update</span>
+                                          </button>
+                                          <button type="submit" class="btn btn-primary ml-1" data-bs-dismiss="modal" name="setPassDef" onclick="return confirmReset()">
+                                            <input type="text" class="visually-hidden" value="<?= $user['id'] ?>" name="id">
+                                            <i class="bx bx-check d-block d-sm-none"></i>
+                                            <span class="d-none d-sm-block">Set Password To Default</span>
                                           </button>
                                         </div>
                                       </form>
                                     </div>
                                   </div>
                                 </div>
-                                <a href="?delete=<?= $user['id'] ?>" class="btn icon icon-left btn-danger"><i class="bi bi-x"></i>
+                                <a href="?delete=<?= $user['id'] ?>" class="btn icon icon-left btn-danger" onclick="return confirm('Apakah Anda yakin ingin menghapus user ini?')"><i class="bi bi-x"></i>
                                   DELETE</a>
                               </td>
                             </tr>
@@ -169,7 +169,6 @@ if (isset($_POST['simpan'])) {
                       </table>
                     </div>
                   </div>
-
                   <div class="card-footer d-flex justify-content-between">
                     <a href="../register.php" class="btn icon icon-left btn-success" data-bs-toggle="modal" data-bs-target="#default"><i class="bi bi-person-plus-fill"></i>
                       TAMBAH</a>
@@ -183,7 +182,6 @@ if (isset($_POST['simpan'])) {
                               <i data-feather="x"></i>
                             </button>
                           </div>
-
                           <form action="" method="post">
                             <div class="modal-body">
                               <label>Nama : </label>
@@ -193,10 +191,6 @@ if (isset($_POST['simpan'])) {
                               <label>Username: </label>
                               <div class=" form-group">
                                 <input type="text" placeholder="Username" class="form-control" name="username" value="">
-                              </div>
-                              <label>Password: </label>
-                              <div class=" form-group">
-                                <input type="password" placeholder="Password" class="form-control" name="password" value="">
                               </div>
                               <label>Role: </label>
                               <div class=" form-group">
@@ -212,13 +206,11 @@ if (isset($_POST['simpan'])) {
                                 <span class="d-none d-sm-block">Close</span>
                               </button>
                               <button type="submit" class="btn btn-primary ml-1" data-bs-dismiss="modal" name="simpan" value="SIMPAN DATA">
-                                <!-- <input type="text" class="visually-hidden" value="" name="id"> -->
                                 <i class="bx bx-check d-block d-sm-none"></i>
                                 <span class="d-none d-sm-block">Simpan</span>
                               </button>
                             </div>
                           </form>
-
                         </div>
                       </div>
                     </div>
@@ -231,11 +223,14 @@ if (isset($_POST['simpan'])) {
       </div>
     </div>
   </div>
-
   <script src="../assets/js/bootstrap.js"></script>
   <script src="../assets/js/app.js"></script>
   <script src="../assets/extensions/simple-datatables/umd/simple-datatables.js"></script>
   <script src="../assets/js/pages/simple-datatables.js"></script>
+  <script>
+    function confirmReset() {
+      return confirm('Apakah Anda yakin ingin mengatur ulang kata sandi ke default pada user ini?');
+    }
+  </script>
 </body>
-
 </html>
